Fixed a bug with idea import function
There was a bug with the way that the import function would handle rows that had a null value in the idea column. Updated the code to check if the idea number column is null for a given row. If it is null, skip it and output a warning to the log.